+++ chatlearn/models/torch_module.py
@@ -88,7 +88,9 @@ class TorchModule(BaseModule):
         """
         :meta private:
         """
+        self._logger.info(f"======================= start getting peak memory at rank {self.rank}")
         self._peak_memory = max(self._peak_memory, torch.cuda.max_memory_allocated() / (1024 ** 3))
+        self._logger.info(f"done getting peak memory at rank {self.rank} =====================")
         return self._peak_memory
 
     def empty_cache(self):
