+++ chatlearn/runtime/executor.py
@@ -19,7 +19,7 @@ from collections import defaultdict
 from itertools import cycle
 from ray.util.queue import Queue
 
-from chatlearn.runtime.model_flow import ModelFlow, ModelNode
+from chatlearn.runtime.model_flow import ModelFlow
 from chatlearn.utils import future
 from chatlearn.utils.global_vars import get_args
 from chatlearn.utils.logger import logger
